requirements.txt
Django==4.2.8
django-cors-headers==4.3.1
djangorestframework==3.14.0
anthropic==0.7.8
python-docx==1.1.0
PyPDF2==3.0.1
pandas>=2.1.0
numpy>=1.26.0
openpyxl==3.1.2
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
markdown==3.5.1
beautifulsoup4==4.12.2
dj-database-url==2.1.0
